integrate gigaget further into newpipe
