+++ app/src/main/java/org/schabi/newpipe/downloadmanager/get/DownloadDataSource.java
@@ -1,8 +1,24 @@
-package us.shandian.giga.get;
+package org.schabi.newpipe.downloadmanager.get;
 
 import java.util.List;
 
 /**
+ * Copyright (C) 2014 Peter Cai
+ * Changes by Christian Schabesberger (C) 2018
+ *
+ * org.schabi.newpipe.downloadmanager is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * org.schabi.newpipe.downloadmanager is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with org.schabi.newpipe.downloadmanager.  If not, see <http://www.gnu.org/licenses/>.
+ *
  * Provides access to the storage of {@link DownloadMission}s
  */
 public interface DownloadDataSource {
