+++ app/src/main/java/org/schabi/newpipe/downloadmanager/get/DownloadManagerImpl.java
@@ -1,4 +1,4 @@
-package us.shandian.giga.get;
+package org.schabi.newpipe.downloadmanager.get;
 
 import android.support.annotation.Nullable;
 import android.util.Log;
@@ -17,9 +17,26 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-import us.shandian.giga.util.Utility;
+import org.schabi.newpipe.downloadmanager.util.Utility;
 
 import static org.schabi.newpipe.BuildConfig.DEBUG;
+/**
+ * Copyright (C) 2014 Peter Cai
+ * Changes by Christian Schabesberger (C) 2018
+ *
+ * org.schabi.newpipe.downloadmanager is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * org.schabi.newpipe.downloadmanager is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with org.schabi.newpipe.downloadmanager.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 public class DownloadManagerImpl implements DownloadManager {
     private static final String TAG = DownloadManagerImpl.class.getSimpleName();
