+++ app/src/main/java/org/schabi/newpipe/downloadmanager/get/DownloadMission.java
@@ -1,4 +1,4 @@
-package us.shandian.giga.get;
+package org.schabi.newpipe.downloadmanager.get;
 
 import android.os.Handler;
 import android.os.Looper;
@@ -14,9 +14,28 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import us.shandian.giga.util.Utility;
+import org.schabi.newpipe.downloadmanager.util.Utility;
 
 import static org.schabi.newpipe.BuildConfig.DEBUG;
+
+
+/**
+ * Copyright (C) 2014 Peter Cai
+ * Changes by Christian Schabesberger (C) 2018
+ *
+ * org.schabi.newpipe.downloadmanager is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * org.schabi.newpipe.downloadmanager is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with org.schabi.newpipe.downloadmanager.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 public class DownloadMission {
     private static final String TAG = DownloadMission.class.getSimpleName();
