+++ app/src/main/java/org/schabi/newpipe/downloadmanager/get/sqlite/DownloadMissionSQLiteHelper.java
@@ -1,4 +1,4 @@
-package us.shandian.giga.get.sqlite;
+package org.schabi.newpipe.downloadmanager.get.sqlite;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -6,10 +6,26 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import us.shandian.giga.get.DownloadMission;
+import org.schabi.newpipe.downloadmanager.get.DownloadMission;
 
 /**
- * SqliteHelper to store {@link us.shandian.giga.get.DownloadMission}
+ * Copyright (C) 2014 Peter Cai
+ * Changes by Christian Schabesberger (C) 2018
+ *
+ * org.schabi.newpipe.downloadmanager is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * org.schabi.newpipe.downloadmanager is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with org.schabi.newpipe.downloadmanager.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * SqliteHelper to store {@link DownloadMission}
  */
 public class DownloadMissionSQLiteHelper extends SQLiteOpenHelper {
