+++ app/src/main/java/org/schabi/newpipe/downloadmanager/ui/common/ProgressDrawable.java
@@ -1,4 +1,4 @@
-package us.shandian.giga.ui.common;
+package org.schabi.newpipe.downloadmanager.ui.common;
 
 import android.content.Context;
 import android.graphics.Canvas;
@@ -9,6 +9,24 @@ import android.graphics.drawable.Drawable;
 import android.support.annotation.ColorRes;
 import android.support.annotation.NonNull;
 import android.support.v4.content.ContextCompat;
+
+/**
+ * Copyright (C) 2014 Peter Cai
+ * Changes by Christian Schabesberger (C) 2018
+ *
+ * org.schabi.newpipe.downloadmanager is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * org.schabi.newpipe.downloadmanager is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with org.schabi.newpipe.downloadmanager.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 public class ProgressDrawable extends Drawable {
     private float mProgress;
